+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Unreleased
 ----------
 
 - Do not remove leading unfilled semicolons in the rendered performance data
+- Add additional parameters (``invert``, ``start``, ``end``) to the constructor of the ``Range`` class
 
 2.0.0 (2026-03-02)
 ------------------